requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
japanize-matplotlib==1.1.3
lightning-utilities==0.10.0
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
pandas==2.1.3
Pillow==10.1.0
python-dateutil==2.8.2
pytorch-lightning==2.1.2
scikit-learn==1.3.2
shap==0.43.0
streamlit==1.28.2
torch==2.1.1
